+++ mtests/TestSuiteDie.c
@@ -5,7 +5,7 @@
 void test_roll_returns_reasonable_result() {
 	Die *d = die_create(6);
 	
-	int result = die_roll(d);
+	long long int result = die_roll(d);
 
 	CU_ASSERT(result >= 1);
 	CU_ASSERT(result <= 6);
@@ -15,7 +15,7 @@ void test_roll_returns_reasonable_result() {
 void test_die_saves_result() {
 	Die *d = die_create(6);
 	
-	size_t result = die_roll(d);
+	long long int result = die_roll(d);
 	
 	CU_ASSERT_EQUAL(result, die_last_result(d));
 }
